Add second storyboard for settings. Make some refactoring
