+++ practice_1/Controllers/SettingsViewController.swift
@@ -1,0 +1,19 @@
+//
+//  SettingsViewController.swift
+//  practice_1
+//
+//  Created by user199993 on 6/24/21.
+//
+
+import UIKit
+
+class SettingsViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+
+    }
+    
+
+}
