+++ practice_1/Utils/RandomNumber.swift
@@ -1,0 +1,5 @@
+class RandomNumber {
+    static func getRandomNumber(_ startOfRange: Int, _ endOfRange: Int) -> Int {
+        Int.random(in: startOfRange...endOfRange)
+    }
+}
